Fix some compiler warnings in the reordered_opaque code.
git-svn-id: http://svn.mythtv.org/svn/trunk@25669 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/avformatdecoder.cpp
@@ -2988,12 +2988,12 @@ bool AvFormatDecoder::ProcessVideoPacket(AVStream *curstream, AVPacket *pkt)
         return false;
     }
 
-    if ((use_reordered_opaque || pkt->dts == AV_NOPTS_VALUE) &&
-        mpa_pic.reordered_opaque != AV_NOPTS_VALUE)
+    if ((use_reordered_opaque || pkt->dts == (int64_t)AV_NOPTS_VALUE) &&
+        mpa_pic.reordered_opaque != (int64_t)AV_NOPTS_VALUE)
     {
         pts = (long long)mpa_pic.reordered_opaque;
     }
-    else if (pkt->dts != AV_NOPTS_VALUE)
+    else if (pkt->dts != (int64_t)AV_NOPTS_VALUE)
     {
         pts = (long long)pkt->dts;
     }
